+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/compose/general/ChatTheme.kt
@@ -147,6 +147,16 @@ private object ChatThemeDateFormatterSnippet : ChatThemeCustomization() {
                     0
                 ).toString()
             }
+
+            override fun formatRelativeDate(date: Date): String {
+                // Provide a way to format Relative Date
+                return DateUtils.getRelativeTimeSpanString(
+                    date.time,
+                    System.currentTimeMillis(),
+                    DateUtils.DAY_IN_MILLIS,
+                    DateUtils.FORMAT_ABBREV_RELATIVE,
+                ).toString()
+            }
         }
     }
 }